+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `WakuRelay.getPeers` method.
+
 ### Changed
 - Enable passing `string`s to `addPeerToAddressBook`.
 - Use `addPeerToAddressBook` in examples and usage doc.